Separate client into Client object
ClientStatus enum is moved to protocol.rs so that Client and Server both
can use it.

Client object works mostly the same as the client-related functions that
used to exist in main.rs, but with the hope that with more work (and
modifications to the protocol), client can request parts of file
concurrently and piece them together to better utilize the network
capacity available to both client and server.