Add new message types in protocol.rs
One is an END message to signal to client that there is now no more data
to send. This is a precaution to avoid having clients wait around for
more data

Another is a function to build a data packet. This data packet will
contain offset data, size of the actual data, and the data within the
constraints of protocol::MTU

This will aid immensely in implementing concurrent client request
support for faster data transfer speeds.