Add NAT traversal
This commit adds code to radically change the way client-server
communication is setup. Upon execution, it connects to 2 separate STUN
servers and checks if IP-port is same for both, indicating easy NAT.

If easy NAT is detected, it will ask for other party's external IP-port
and attempt to send ACKs to the other party in order to open up the
firewalls temporarily for file transfer. The other machine is doing the
same thing. In order to ensure that this happens around the same time, a
buffer of 5 seconds is maintained before and after this procedure, after
which the program proceeds pretty much as before.

More work will likely be needed server side in order to have multiple
users connecting to same machine, however this is an excellent start!