Add preliminary support for increasing window size
Min value is 6 and maximum is 75. The window size changing relies on a
Mutex which does complicate things, and is increased and decreased as
follows:

Increased by 1 every time N packets have come through successfully
Decreased to 1 when a resend has been sent by the other party

While theoretical max is 255, there were some errors while testing on
localhost, so for now it is set to 75, which seemed to have worked. This
commit is experimental, and more work WILL be needed to actually have
this feature fully implemented