Migrate main loop code of server to server.rs
This is to later better handle selective repeat requests, and a similar
operation will be performed for client.rs for symmetry and to handle
similar cases for client in case we need to in the future

Server was further improved by removing unnecessary task spawns in
process_msg(), as these did not yield great benefit by being run in
another thread, and, due to the usage of Mutexes, would actually incur a
penalty instead (based on some basic testing). The only task spawning is
done when server has to send multiple packets at a time, which makes
more sense to be sped up using concurrency.